exercise_8/requirements.txt
# API requests
requests>=2.31.0

# Data manipulation and analysis
pandas>=2.1.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Environment variables
python-dotenv>=1.0.0

# Type hints support
typing-extensions>=4.7.0

# Date handling
pytz>=2023.3

# Optional: Jupyter notebook support (if you want to run analysis interactively)
jupyter>=1.0.0
ipykernel>=6.25.0